# FIX: Domain entity, and slow __repr__