Fixes #557 and #555 

- Uses only the presence of 'derivatives' in config to determine if a `BIDSLayout` is in the `derivative` domain. I could see an argument that we should consider all which have more than only the core `bids` config listed.

- I used some very inelegant (rather janky) counts using just the `Tag` model as Tal suggested. 
%timeit results using the `synthetic` test dataset:

Previous method: 606ms
This PR: 4.99ms
Only counting distinct session id's, not nested within subjects: 3.99ms (so not worth only doing this)

I will say that just doing `layout.get_subjects` takes 35ms on this dataset, so there's room for optimization on `.get`. 

On a real dataset with no sessions (life), I saw a 10x speedup: 2.83s vs 0.27s