@adelavega 2020-01-08
On ds003_fmriprep, initated with derivatives= same directory and validate=False, I was getting extremely slow performance before (`308 s`) and with this PR it only took `0.013` seconds...

yuuuge, if I may say so myself.