@adelavega 2020-01-08
Okay, there is one potential problem here, my method is only looking at tags in the top level layout, whereas previously it was getting unique combinations across levels (perhaps part of the reason it was extremely slow at times)

In this case:
`BIDSLayout('fmirprep_dir', derivatives='fmirprep_dir')`

`__repr__` would return no subjects because no files passed the `bids` config in the top level `BIDSLayout`.

I could tweak my ugly hacks to also loop overall tags in derivatives and that should probably fix it, but I wonder if its worth the peformance cost?