@adelavega 2020-01-10
@tyarkoni as is, it only computes this for the top level layout, not any derivatives. is that okay? I can have it join all of the tags from all the children's layouts to get a complete result.

if that sounds good, ill add that + chris' commments , and merge if tests pass,