@tyarkoni 2020-01-10
Fine with me—I think that's actually the behavior I would expect.